Reader: Stash the last and first date in the stream instead of looking them up.
With the previous method, when a stream had a lot of x-posts in it, there was a good chance we could end up with the 'last' post in a page of stream posts as an x-post. When we then looked up the last date, we would end up using the wrong date and just loop, over and over again, trying to find more posts. This fixes that by stashing the extents of the stream page on the stream instead of using the posts themselves.

Ideally we'll switch to using a paging handle, which can encode both the date _and_ an id for the post (likely feed id / item id or site id / post id) and we'll switch to stashing those instead.