Domains: Design for domain credit with DWPO
Currently, when the user is in plansOnly group
and have domain credit, they see the a-la-carte
domain pricing scratched off and says "Free with your plan".
However, they were not able to buy their first domain
a-la-carte anyway. I think it might a good improvement
to change the presentation that doesn't mention
a-la-carte prices.